fix: make sanitizeInstrumentUnit a private MeterImpl method
Address review feedback on #932: move sanitizeInstrumentUnit off the
SdkErrorHandler extension namespace and into a private MeterImpl
method, since it's an implementation detail other classes shouldn't
see. Test coverage moves from the standalone InstrumentUnitTest into
UpDownCounterImplTest, exercising sanitization through the public
Meter API.
